feat: entitiy 수정중 WIP

+++ BackEnd/src/main/java/com/minePing/BackEnd/entity/Health.java
@@ -23,7 +23,7 @@ public class Health {
     private Member member;
 
     @Column(name="weight",nullable = false)
-    private Integer weight;
+    private Float weight;
 
     @Column(name="height",nullable = false)
     private Float height;
